refactor: screen フォルダを作成
