+++ app/src/main/java/com/rnk0085/selfcare/App.kt
@@ -5,7 +5,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import com.rnk0085.selfcare.ui.screen.Screen
 import com.rnk0085.selfcare.ui.screen.diaryList.diaryList
-import com.rnk0085.selfcare.ui.screen.reflection.reflection
+import com.rnk0085.selfcare.ui.screen.reflection.screen.reflection
 
 @Composable
 internal fun App() {
