+++ app/src/main/java/com/rnk0085/selfcare/ui/screen/reflection/screen/ReflectionScreen.kt
@@ -1,4 +1,4 @@
-package com.rnk0085.selfcare.ui.screen.reflection
+package com.rnk0085.selfcare.ui.screen.reflection.screen
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
